minor fix & support other prompting methods (#1)
* minor fix

* support other prompting methods

* add paper link

+++ train.py
@@ -8,6 +8,7 @@ from dassl.utils import setup_logger, set_random_seed, collect_env_info
 from dassl.config import get_cfg_default
 from dassl.engine import build_trainer
 import os
+from utils import print_args, is_main_process
 # custom
 import datasets.oxford_pets
 import datasets.oxford_flowers
@@ -28,20 +29,11 @@ import datasets.imagenet_r
 
 import trainers.pomp
 import trainers.clip_mlp
-
-
-def print_args(args, cfg):
-    print("***************")
-    print("** Arguments **")
-    print("***************")
-    optkeys = list(args.__dict__.keys())
-    optkeys.sort()
-    for key in optkeys:
-        print("{}: {}".format(key, args.__dict__[key]))
-    print("************")
-    print("** Config **")
-    print("************")
-    print(cfg)
+import trainers.coop
+import trainers.cocoop
+import trainers.zsclip
+import trainers.maple
+import trainers.vpt
 
 
 def reset_cfg(cfg, args):
@@ -92,6 +84,9 @@ def extend_cfg(cfg):
     """
     from yacs.config import CfgNode as CN
 
+    cfg.DATASET.SUBSAMPLE_CLASSES = "all"  # all, base or new
+
+    # Config for POMP
     cfg.TRAINER.POMP = CN()
     cfg.TRAINER.POMP.N_CTX = 4  # number of context vectors
     cfg.TRAINER.POMP.CSC = False  # class-specific context
@@ -103,6 +98,36 @@ def extend_cfg(cfg):
     cfg.TRAINER.POMP.LOCAL_CORRECTION = True
     cfg.TRAINER.POMP.NEG_SAMPLING_MODE = 'random'
 
+    # Config for COOP
+    cfg.TRAINER.COOP = CN()
+    cfg.TRAINER.COOP.N_CTX = 16  # number of context vectors
+    cfg.TRAINER.COOP.CSC = False  # class-specific context
+    cfg.TRAINER.COOP.CTX_INIT = ""  # initialization words
+    cfg.TRAINER.COOP.PREC = "fp16"  # fp16, fp32, amp
+    cfg.TRAINER.COOP.CLASS_TOKEN_POSITION = "end"  # 'middle' or 'end' or 'front'
+    cfg.TRAINER.COOP.PRETRAINED_WEOGHT = ''
+
+    # Config for COCOOP
+    cfg.TRAINER.COCOOP = CN()
+    cfg.TRAINER.COCOOP.N_CTX = 16  # number of context vectors
+    cfg.TRAINER.COCOOP.CTX_INIT = ""  # initialization words
+    cfg.TRAINER.COCOOP.PREC = "fp16"  # fp16, fp32, amp
+
+    # Config for MaPLe
+    cfg.TRAINER.MAPLE = CN()
+    cfg.TRAINER.MAPLE.N_CTX = 2  # number of context vectors
+    cfg.TRAINER.MAPLE.CTX_INIT = "a photo of a"  # initialization words
+    cfg.TRAINER.MAPLE.PREC = "fp16"  # fp16, fp32, amp
+    cfg.TRAINER.MAPLE.PROMPT_DEPTH = 9 # Max 12, minimum 0, for 1 it will act as shallow MaPLe (J=1)
+
+    # Config for only vision side prompting
+    cfg.TRAINER.VPT = CN()
+    cfg.TRAINER.VPT.N_CTX_VISION = 2  # number of context vectors at the vision branch
+    cfg.TRAINER.VPT.CTX_INIT = "a photo of a"  # initialization words
+    cfg.TRAINER.VPT.PREC = "fp16"  # fp16, fp32, amp
+    cfg.TRAINER.VPT.PROMPT_DEPTH_VISION = 1  # if set to 1, will represent shallow vision prompting only
+
+    # Config for MLP & Linear
     cfg.TRAINER.MLP = CN()
     cfg.TRAINER.MLP.PREC = "fp16"  # fp16, fp32, amp
 
@@ -142,15 +167,7 @@ def main(args):
         dist.init_process_group(backend="nccl", world_size=args.world_size, rank=args.local_rank)
         torch.cuda.set_device(args.local_rank)
 
-    if dist.is_initialized():
-        if cfg.SEED >= 0:
-            print("Setting fixed seed: {}".format(cfg.SEED))
-            set_random_seed(cfg.SEED + dist.get_rank())
-        if dist.get_rank() == 0:
-            print_args(args, cfg)
-            print("Collecting env info ...")
-            print("** System info **\n{}\n".format(collect_env_info()))
-    else:
+    if is_main_process():
         if cfg.SEED >= 0:
             print("Setting fixed seed: {}".format(cfg.SEED))
             set_random_seed(cfg.SEED)
@@ -182,7 +199,9 @@ def main(args):
         best_checkpoint_file = "model.pth.tar-" + str(best_checkpoint_id)
         best_checkpoint_path = os.path.join(args.model_dir, learner, best_checkpoint_file)
         rename_path = os.path.join(args.model_dir, learner, 'model-best.pth.tar')
-        os.system(f"mv {best_checkpoint_path} {rename_path}")
+        if is_main_process():
+            os.system(f"mv {best_checkpoint_path} {rename_path}")
+        dist.barrier()
 
         trainer.load_model(args.model_dir, epoch=None)
         trainer.test(split='test')
@@ -191,7 +210,8 @@ def main(args):
             if epoch != best_checkpoint_id:
                 checkpoint_file = "model.pth.tar-" + str(epoch)
                 checkpoint_path = os.path.join(args.model_dir, learner, checkpoint_file)
-                os.system(f"rm -rf {checkpoint_path}")
+                if is_main_process():
+                    os.system(f"rm -rf {checkpoint_path}")
         return
     if not args.no_train:
         if args.model_dir != '':
